+++ data-transfer-object/src/main/java/org/grnet/cat/dtos/assessment/AssessmentResponse.java
@@ -1,13 +1,10 @@
-package org.grnet.cat.dtos;
-
+package org.grnet.cat.dtos.assessment;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.eclipse.microprofile.openapi.annotations.enums.SchemaType;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
-import org.json.simple.JSONObject;
 
-@Schema(name = "Assessment", description = "This object represents the Assessment.")
-public class AssessmentResponseDto {
+public class AssessmentResponse {
 
     @Schema(
             type = SchemaType.NUMBER,
@@ -21,23 +18,9 @@ public class AssessmentResponseDto {
     @Schema(
             type = SchemaType.NUMBER,
             implementation = Long.class,
-            description = "The template id for the assessment")
-    @JsonProperty("templateId")
-    public Long templateId;
-
-    @Schema(
-            type = SchemaType.NUMBER,
-            implementation = Long.class,
             description = "The validation id of the assessment")
     @JsonProperty("validationId")
     public Long validationId;
-
-    @Schema(
-            type = SchemaType.OBJECT,
-            implementation = JSONObject.class,
-            description = "The assessment doc")
-    @JsonProperty("assessmentDoc")
-    public JSONObject assessmentDoc;
 
     @Schema(
             type = SchemaType.STRING,